requirements.txt
# Core GUI Framework
PySide6>=6.5.0

# Database ORM
SQLAlchemy>=1.4.0

# Image Processing & OCR
opencv-python>=4.5.0
Pillow>=8.0.0
pytesseract>=0.3.0
pdf2image>=1.16.0

# Enhanced OCR Engines (choose one or both)
paddleocr>=2.7.0  # Recommended: Best accuracy, supports CPU/GPU
# easyocr>=1.7.0  # Alternative: Lightweight, good for English

# Table Detection & AI Models
torch>=2.0.0  # PyTorch for Table Transformer
torchvision>=0.15.0
transformers>=4.30.0  # Hugging Face Transformers for Table Transformer model
timm>=0.9.0  # Required by Table Transformer

# Optional: Cloud OCR APIs (uncomment if needed)
# google-cloud-vision>=3.4.0  # Google Vision API
# boto3>=1.28.0  # AWS Textract
# azure-ai-formrecognizer>=3.3.0  # Azure Document Intelligence

# PDF Processing
PyMuPDF>=1.21.0
pdfplumber>=0.10.0

# Excel Export
openpyxl>=3.0.10

# Scientific Computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Plotting and Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Date/Time utilities
python-dateutil>=2.8.0

# Markdown parsing for help system
markdown>=3.4.0

# Development dependencies (optional)
pytest>=7.0.0